test(sdk): share compiler work across public contract checks
