+++ packages/sdk/test/public-contracts.test.ts
@@ -2,16 +2,19 @@ import { fileURLToPath } from 'node:url';
 import ts from 'typescript';
 import { describe, expect, it } from 'vitest';
 import type {
+  AwsConfigRecordingFrequencyReview,
   AwsCostOptimizationHubAutoScalingUpgradeConfiguration,
   AwsCostOptimizationHubEbsUpgradeConfiguration,
   AwsCostOptimizationHubEc2UpgradeConfiguration,
   AwsCostOptimizationHubRdsStorageUpgradeConfiguration,
   AwsCostOptimizationHubRdsUpgradeConfiguration,
   AwsCostOptimizationHubUpgradeRecommendation,
+  LiveEvaluationCoverage,
+  RuleEvaluation,
 } from '../src/index.js';
 
-describe('public upgrade configuration contracts', () => {
-  it('exports typed configurations from the SDK entry point and type module', () => {
+describe('public SDK contracts', () => {
+  it('exports compiler-checked upgrade configurations, evaluation coverage and Config evidence', () => {
     const configPath = fileURLToPath(new URL('../../../tsconfig.base.json', import.meta.url));
     const config = ts.readConfigFile(configPath, ts.sys.readFile);
     const parsed = ts.parseJsonConfigFileContent(
@@ -41,5 +44,23 @@ describe('public upgrade configuration contracts', () => {
     const type: AwsCostOptimizationHubUpgradeRecommendation['actionType'] = 'Upgrade';
     expect(type).toBe('Upgrade');
     expect(shapes).toHaveLength(5);
-  }, 15_000);
+
+    const coverage: LiveEvaluationCoverage = { assessed: [], unknown: [{ resourceId: 'function' }] };
+    const evaluation: Pick<RuleEvaluation, 'status' | 'coverage'> = { coverage, status: 'unknown' };
+    const evidence: Pick<
+      AwsConfigRecordingFrequencyReview,
+      | 'configurationItemsRecorded'
+      | 'estimatedMonthlyConfigurationItemReduction'
+      | 'estimatedMonthlyRecordingCostReductionUsd'
+    > = {
+      configurationItemsRecorded: null,
+      estimatedMonthlyConfigurationItemReduction: null,
+      estimatedMonthlyRecordingCostReductionUsd: null,
+    };
+    // @ts-expect-error Unknown metric evidence must be checked before arithmetic.
+    const volume: number = evidence.configurationItemsRecorded;
+    // @ts-expect-error Coverage identities require a resource ID.
+    const invalidCoverage: LiveEvaluationCoverage = { assessed: [{}], unknown: [] };
+    void [evaluation, volume, invalidCoverage];
+  }, 30_000);
 });
